requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.0.6
django-ckeditor==6.7.1
django-cors-headers==4.4.0
django-js-asset==2.2.0
djangorestframework==3.15.2
gunicorn==22.0.0
mysqlclient==2.2.4
packaging==24.1
psycopg2-binary==2.9.9
sqlparse==0.5.0
typing_extensions==4.12.2
